refactor: 딜러와 플레이어의 어색한 패키지 위치 수정
